♻️ Extract sub-component for Account
Rename AccountLive -> HomeLive and extract new AccountLive as a live component.

HomeLive loads the account on mount and controls the page title.

AccountLive renders the account and handles its editing.

FundsLive continues to render the fund list and control their editing.
