+++ lib/freedom_account_web/live/home_live/show.html.heex
@@ -1,0 +1,14 @@
+<.live_component
+  account={@account}
+  action={@live_action}
+  id={@account.id}
+  module={FreedomAccountWeb.AccountLive.Show}
+  title={@page_title}
+/>
+<.live_component
+  account={@account}
+  action={@live_action}
+  id={@account.id}
+  module={FreedomAccountWeb.FundLive.Index}
+  title={@page_title}
+/>
